+++ .changes/rc-migration.md
@@ -1,0 +1,6 @@
+---
+"tauri-cli": patch:feat
+"@tauri-apps/cli": patch:feat
+---
+
+Added migration from `2.0.0-beta` to `2.0.0-rc`.
