+++ tooling/cli/Cargo.toml
@@ -98,6 +98,7 @@ oxc_allocator = "0.16"
 oxc_ast = "0.16"
 magic_string = "0.3"
 phf = { version = "0.11", features = ["macros"] }
+walkdir = "2"
 
 [target."cfg(windows)".dependencies.windows-sys]
 version = "0.52"
